feat(web): 모달 레이아웃 및 디자인 수정
